+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "axios": "^0.21.1",
     "bootstrap": "^4.5.0",
+    "i": "^0.3.6",
     "jquery": "^3.5.1",
     "popper.js": "^1.16.1"
   }
